+++ src/lib.rs
@@ -5,6 +5,7 @@
 pub mod config;
 pub mod dependency;
 pub mod easysetup;
+pub mod http;
 pub mod mqtt;
 pub mod provisioning;
 pub mod util;
